Moved wildcard matching logic into util functions
